requirements.txt
faster_whisper
ffmpeg-python
loguru
PyYAML
